refactor: separate model to sub-dirs
