+++ src/main/kotlin/blackjack/model/card/Suit.kt
@@ -1,4 +1,4 @@
-package blackjack.model
+package blackjack.model.card
 
 enum class Suit(val symbol: String) {
     HEART("♥"),
