+++ src/main/kotlin/blackjack/model/participant/PlayerFactory.kt
@@ -1,4 +1,4 @@
-package blackjack.model
+package blackjack.model.participant
 
 object PlayerFactory {
     fun createPlayers(names: List<String>) = names.map { Player(it) }
