+++ src/test/kotlin/blackjack/model/CardDeckTest.kt
@@ -1,5 +1,7 @@
 package blackjack.model
 
+import blackjack.model.game.CardDeck
+import blackjack.model.participant.Player
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
