+++ src/test/kotlin/blackjack/model/PlayerFactoryTest.kt
@@ -1,5 +1,6 @@
 package blackjack.model
 
+import blackjack.model.participant.PlayerFactory
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
